+++ sds/src/lib.rs
@@ -2,7 +2,6 @@
 // override with `#[allow(clippy::print_stdout)]`.
 #![deny(clippy::print_stdout)]
 
-mod config;
 mod encoding;
 mod event;
 mod match_action;
@@ -29,9 +28,11 @@ pub use path::{Path, PathSegment};
 pub use rule_match::{ReplacementType, RuleMatch};
 pub use scanner::cache_pool::{CachePool, CachePoolBuilder, CachePoolGuard};
 pub use scanner::{
-    config::ProximityKeywordsConfig, config::RuleConfig, error::CreateScannerError,
-    regex_rule::config::RegexRuleConfig, scope::Scope, CompiledRuleTrait, MatchEmitter, Scanner,
-    ScannerBuilder, StringMatch,
+    config::RuleConfig,
+    error::CreateScannerError,
+    regex_rule::config::{ProximityKeywordsConfig, RegexRuleConfig},
+    scope::Scope,
+    CompiledRuleTrait, MatchEmitter, Scanner, ScannerBuilder, StringMatch,
 };
 pub use scoped_ruleset::ExclusionCheck;
 pub use validation::{
