+++ sds/src/proximity_keywords/included_keywords.rs
@@ -56,7 +56,7 @@ mod test {
     use crate::proximity_keywords::{
         compile_keywords_proximity_config, CompiledIncludedProximityKeywords,
     };
-    use crate::scanner::config::ProximityKeywordsConfig;
+    use crate::scanner::regex_rule::config::ProximityKeywordsConfig;
     use crate::Labels;
 
     fn compile_keywords(lookahead: usize, keywords: &[&str]) -> CompiledIncludedProximityKeywords {
