+++ sds/src/scanner/mod.rs
@@ -10,7 +10,6 @@ use std::sync::Arc;
 
 use self::cache_pool::{CachePool, CachePoolBuilder, CachePoolGuard};
 use self::metrics::ScannerMetrics;
-// use crate::scanner::regex_rule::RegexCompiledRule;
 use crate::scanner::config::RuleConfig;
 use crate::scanner::regex_rule::compiled::RegexCompiledRule;
 use crate::scanner::scope::Scope;
@@ -465,9 +464,7 @@ mod test {
     use super::{MatchEmitter, ScannerBuilder, StringMatch};
     use crate::match_action::{MatchAction, MatchActionValidationError};
     use crate::observability::labels::Labels;
-    use crate::scanner::config::SecondaryValidator::*;
-    use crate::scanner::config::{ProximityKeywordsConfig, SecondaryValidator};
-    use crate::scanner::regex_rule::config::RegexRuleConfig;
+    use crate::scanner::regex_rule::config::{ProximityKeywordsConfig, SecondaryValidator, SecondaryValidator::*, RegexRuleConfig};
     use crate::scanner::scope::Scope;
     use crate::scanner::{get_next_regex_start, CreateScannerError, Scanner};
     use crate::scoped_ruleset::ExclusionCheck;
@@ -1803,8 +1800,7 @@ mod test {
 
     mod metrics_test {
         use crate::match_action::MatchAction;
-        use crate::scanner::config::ProximityKeywordsConfig;
-        use crate::scanner::regex_rule::config::RegexRuleConfig;
+        use crate::scanner::regex_rule::config::{RegexRuleConfig, ProximityKeywordsConfig};
         use crate::scanner::scope::Scope;
         use crate::scanner::ScannerBuilder;
         use crate::{simple_event::SimpleEvent, Path, PathSegment};
