+++ sds/src/scanner/regex_rule/config.rs
@@ -1,5 +1,5 @@
 use crate::proximity_keywords::compile_keywords_proximity_config;
-use crate::scanner::config::{ProximityKeywordsConfig, RuleConfig, SecondaryValidator};
+use crate::scanner::config::RuleConfig;
 use crate::scanner::metrics::RuleMetrics;
 use crate::scanner::regex_rule::compiled::RegexCompiledRule;
 use crate::scanner::scope::Scope;
@@ -113,6 +113,30 @@ impl RuleConfig for RegexRuleConfig {
     }
 }
 
+#[serde_as]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
+pub struct ProximityKeywordsConfig {
+    pub look_ahead_character_count: usize,
+
+    #[serde_as(deserialize_as = "DefaultOnNull")]
+    #[serde(default)]
+    pub included_keywords: Vec<String>,
+
+    #[serde_as(deserialize_as = "DefaultOnNull")]
+    #[serde(default)]
+    pub excluded_keywords: Vec<String>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
+#[serde(tag = "type")]
+pub enum SecondaryValidator {
+    LuhnChecksum,
+    ChineseIdChecksum,
+    GithubTokenChecksum,
+    NhsCheckDigit,
+    IbanChecker,
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
@@ -135,6 +159,31 @@ mod test {
                 proximity_keywords: None,
                 validator: None,
                 labels: Labels::empty(),
+            }
+        );
+    }
+
+    #[test]
+    fn proximity_keywords_should_have_default() {
+        let json_config = r#"{"look_ahead_character_count": 0}"#;
+        let test: ProximityKeywordsConfig = serde_json::from_str(json_config).unwrap();
+        assert_eq!(
+            test,
+            ProximityKeywordsConfig {
+                look_ahead_character_count: 0,
+                included_keywords: vec![],
+                excluded_keywords: vec![]
+            }
+        );
+
+        let json_config = r#"{"look_ahead_character_count": 0, "excluded_keywords": null, "included_keywords": null}"#;
+        let test: ProximityKeywordsConfig = serde_json::from_str(json_config).unwrap();
+        assert_eq!(
+            test,
+            ProximityKeywordsConfig {
+                look_ahead_character_count: 0,
+                included_keywords: vec![],
+                excluded_keywords: vec![]
             }
         );
     }
