+++ sds/src/scanner/regex_rule/mod.rs
@@ -1,29 +1,2 @@
 pub mod compiled;
 pub mod config;
-//
-// use super::Validator;
-// use crate::proximity_keywords::{
-//     contains_keyword_in_path, get_prefix_start, is_index_within_prefix,
-//     CompiledExcludedProximityKeywords, CompiledIncludedProximityKeywords,
-// };
-// use crate::scanner::metrics::RuleMetrics;
-// use crate::scanner::{get_next_regex_start, is_false_positive_match};
-// use crate::{
-//     CachePoolGuard, CompiledRuleTrait, ExclusionCheck, MatchAction, MatchEmitter, Path, Scope,
-//     StringMatch,
-// };
-// use ahash::AHashSet;
-// use regex_automata::meta::Cache;
-// use regex_automata::meta::Regex as MetaRegex;
-// use regex_automata::Input;
-// use std::sync::Arc;
-//
-// use std::sync::Arc;
-// use serde::{Deserialize, Serialize};
-// use serde_with::serde_as;
-// use crate::{CachePoolBuilder, CompiledRuleTrait, CreateScannerError, Labels, MatchAction, ProximityKeywordsConfig, RuleConfig, SecondaryValidator};
-// use crate::config::scope::Scope;
-// use crate::proximity_keywords::compile_keywords_proximity_config;
-// use crate::scanner::metrics::RuleMetrics;
-// use crate::scanner::regex_rule::RegexCompiledRule;
-// use crate::validation::validate_and_create_regex;
